the installer is now flexible
