+++ src/future_tstrings/_autoinstall.py
@@ -1,3 +1,3 @@
 from .installer import install
 
-install()
+install(use_import_hook=True)
